Pull StaticChatMessage / MaskedChatMessage into own files
